fix: made it harder to enable vulnerable log statements
Previously, you could enable debug logs and get debug information such
as the contents of canonical requests, JWTs, and the short-lived signing
keys from privy_hpke. This means that someone with the ability to
influence the log level AND view the resulting logs would be able to see
confidential material. This change ensures that the code paths cannot
make it into the binary at compile time, unless you are both running a
debug build AND have the unsafe_debug feature flag enabled.